Extract Cassandra test server runnerr
