+++ testkit/scaladsl/src/main/scala/com/lightbend/lagom/scaladsl/testkit/ServiceTest.scala
@@ -3,17 +3,12 @@
  */
 package com.lightbend.lagom.scaladsl.testkit
 
-import java.nio.file.Files
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
-import java.util.concurrent.TimeUnit
 
-import akka.persistence.cassandra.testkit.CassandraLauncher
-import com.google.common.io.{ MoreFiles, RecursiveDeleteOption }
-import com.lightbend.lagom.internal.testkit.TestConfig
+import com.lightbend.lagom.internal.testkit.{ CassandraTestServer, TestConfig }
 import com.lightbend.lagom.scaladsl.persistence.cassandra.testkit.TestUtil
 import com.lightbend.lagom.scaladsl.server.{ LagomApplication, LagomApplicationContext, RequiresLagomServicePort }
-import org.slf4j.LoggerFactory
 import play.api.ApplicationLoader.Context
 import play.api.inject.DefaultApplicationLifecycle
 import play.api.{ Configuration, Environment, Play }
@@ -37,10 +32,6 @@ import scala.util.control.NonFatal
  * before all tests hook, and then stop it in an after all test hook.
  */
 object ServiceTest {
-
-  private val LagomTestConfigResource: String = "lagom-test-embedded-cassandra.yaml"
-
-  private lazy val log = LoggerFactory.getLogger(getClass)
 
   sealed trait Setup {
     /**
@@ -249,9 +240,11 @@ object ServiceTest {
 
     val config =
       if (setup.cassandra) {
-        configureCassandra(testName, lifecycle)
+        val cassandraPort = CassandraTestServer.run(testName, lifecycle)
+
+        ClusterConfiguration ++ Configuration(TestUtil.persistenceConfig(testName, cassandraPort))
       } else if (setup.jdbc) {
-        JdbcConfiguration ++ ClusterConfiguration
+        ClusterConfiguration ++ JdbcConfiguration
       } else if (setup.cluster) {
         ClusterConfiguration
       } else {
@@ -286,39 +279,6 @@ object ServiceTest {
     }
 
     new TestServer[T](lagomApplication, server)
-  }
-
-  private def configureCassandra[T <: LagomApplication](
-    testName:  String,
-    lifecycle: DefaultApplicationLifecycle
-  ): Configuration = {
-    val cassandraPort = CassandraLauncher.randomPort
-    val cassandraDirectory = Files.createTempDirectory(testName)
-
-    // Shut down Cassandra and delete its temporary directory when the application shuts down
-    lifecycle.addStopHook { () =>
-      import scala.concurrent.ExecutionContext.Implicits.global
-      Try(CassandraLauncher.stop())
-      // The ALLOW_INSECURE option is required to remove the files on OSes that don't support SecureDirectoryStream
-      // See http://google.github.io/guava/releases/snapshot-jre/api/docs/com/google/common/io/MoreFiles.html#deleteRecursively-java.nio.file.Path-com.google.common.io.RecursiveDeleteOption...-
-      Future(MoreFiles.deleteRecursively(cassandraDirectory, RecursiveDeleteOption.ALLOW_INSECURE))
-    }
-
-    val t0 = System.nanoTime()
-
-    CassandraLauncher.start(
-      cassandraDirectory.toFile,
-      LagomTestConfigResource,
-      clean = false,
-      port = cassandraPort,
-      CassandraLauncher.classpathForResources(LagomTestConfigResource)
-    )
-
-    log.debug(s"Cassandra started in ${TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - t0)} ms")
-
-    Configuration(
-      TestUtil.persistenceConfig(testName, cassandraPort, useServiceLocator = false)
-    ) ++ ClusterConfiguration
   }
 
   private lazy val JdbcConfiguration = Configuration(TestConfig.JdbcConfig)
